third_party/WebKit/LayoutTests/fast/dom/Document/createElement-invalid-names-expected.txt

Test various valid name for elements according to the definition of DOM Level 2.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElement('0ascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('.Ascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('-Ascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement(' ascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('	ascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('֑sc.ii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('⃣scii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('िascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('ʳascii') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('asc i') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('asc	i') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS document.createElement('asciiⅦ') threw exception InvalidCharacterError: An invalid or illegal character was specified, such as in an XML name..
PASS successfullyParsed is true

TEST COMPLETE
